+++ src/main/resources/db/migration/prod/V1778187108__Roles_Activities_Values.sql
@@ -1,0 +1,4 @@
+insert into activities values('get operation status');
+
+insert into roles_activities values('admin', 'get operation status');
+insert into roles_activities values('service', 'get operation status');