+++ src/test/kotlin/acceptance/MockConfig.kt
@@ -4,8 +4,10 @@ import io.ktor.http.*
 
 object MockConfig {
     var blizzardProfileStatusOverride: HttpStatusCode? = null
+    var raiderIoCutoffStatusOverride: HttpStatusCode? = null
 
     fun reset() {
         blizzardProfileStatusOverride = null
+        raiderIoCutoffStatusOverride = null
     }
 }